Make sure H.263 plug in code uses size_t for all memory indexing. In particular, fixes a nasty bug where a uin16_t was used, which overflowed when CIF4 is used, causing mayhem.
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_10@27477 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ plugins/video/H.263-1998/rfc2190.h
@@ -48,13 +48,13 @@ class RFC2190Packetizer : public Packetizer
     size_t m_bufferLen;
 
     unsigned int TR;
-    unsigned int frameSize;
+    size_t frameSize;
     int iFrame;
     int annexD, annexE, annexF, annexG, pQuant, cpm;
     int macroblocksPerGOB;
 
     struct fragment {
-      unsigned length;
+      size_t length;
       unsigned mbNum;
     };
 
@@ -64,7 +64,7 @@ class RFC2190Packetizer : public Packetizer
     unsigned char * fragPtr;
 
     unsigned m_currentMB;
-    unsigned m_currentBytes;
+    size_t m_currentBytes;
 };
 
 class RFC2190Depacketizer : public Depacketizer
